# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentJobs API",
    "description": "Search 14,000+ jobs across all industries. Built for AI agents.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://agentjobs.vercel.app",
      "description": "Production"
    }
  ],
  "paths": {
    "/api/v1/jobs": {
      "get": {
        "operationId": "searchJobs",
        "summary": "Search for jobs",
        "description": "Full-text search with filters by role, industry, region, and remote status. Returns freshness-scored results with pagination.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Free-text search query (matches title, company, description)",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "role",
            "in": "query",
            "description": "Role taxonomy slug to filter by (e.g. 'software-engineer')",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "industry",
            "in": "query",
            "description": "Industry slug to filter by (e.g. 'fintech')",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "remote",
            "in": "query",
            "description": "Filter for remote jobs only (true or false)",
            "required": false,
            "schema": { "type": "string", "enum": ["true", "false"] }
          },
          {
            "name": "region",
            "in": "query",
            "description": "Region filter (e.g. US, EU, APAC, LATAM, Global)",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "posted_after",
            "in": "query",
            "description": "ISO 8601 date string; only return jobs posted on or after this date",
            "required": false,
            "schema": { "type": "string", "format": "date-time" }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number (1-indexed)",
            "required": false,
            "schema": { "type": "integer", "minimum": 1, "default": 1 }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Results per page (max 100)",
            "required": false,
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort order for results",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["posted", "freshness", "relevance"],
              "default": "posted"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of matching jobs",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/JobSearchResponse" },
                "example": {
                  "jobs": [
                    {
                      "id": "b1c2d3e4-f5a6-7890-abcd-ef1234567890",
                      "title": "Senior Software Engineer",
                      "company": "Acme Corp",
                      "location": "San Francisco, CA",
                      "link": "https://acme.com/jobs/123",
                      "source": "careers_page",
                      "description": "Build scalable distributed systems...",
                      "salary": "$180k - $220k",
                      "isRemote": false,
                      "region": "US",
                      "postedAt": "2026-03-20T10:00:00.000Z",
                      "lastSeenAt": "2026-03-21T08:00:00.000Z",
                      "isFeatured": false,
                      "freshness": {
                        "score": 0.95,
                        "label": "verified 2 hours ago",
                        "tier": "very_fresh"
                      },
                      "industry": {
                        "id": "uuid",
                        "slug": "technology",
                        "name": "Technology"
                      },
                      "taxonomy": {
                        "id": "uuid",
                        "slug": "software-engineer",
                        "canonicalTitle": "Software Engineer"
                      }
                    }
                  ],
                  "pagination": {
                    "page": 1,
                    "per_page": 20,
                    "total": 342,
                    "total_pages": 18,
                    "has_next": true,
                    "has_prev": false
                  },
                  "meta": {
                    "query_expanded_to": "software engineer",
                    "sort": "posted",
                    "filters": {
                      "role": null,
                      "industry": "technology",
                      "remote": null,
                      "region": null,
                      "posted_after": null
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Optional API key for higher rate limits. Anonymous access is allowed with IP-based rate limiting (10 requests/hour)."
      }
    },
    "schemas": {
      "Freshness": {
        "type": "object",
        "properties": {
          "score": {
            "type": "number",
            "description": "Normalized freshness score from 0 (stale) to 1 (very fresh)"
          },
          "label": {
            "type": "string",
            "description": "Human-readable label (e.g. 'verified 2 hours ago')"
          },
          "tier": {
            "type": "string",
            "enum": ["very_fresh", "fresh", "recent", "aging", "stale", "expired"],
            "description": "Freshness tier category"
          }
        },
        "required": ["score", "label", "tier"]
      },
      "Job": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "company": { "type": "string" },
          "location": { "type": ["string", "null"] },
          "link": { "type": ["string", "null"] },
          "source": { "type": ["string", "null"] },
          "description": { "type": ["string", "null"] },
          "salary": { "type": ["string", "null"] },
          "isRemote": { "type": "boolean" },
          "region": { "type": ["string", "null"] },
          "postedAt": { "type": ["string", "null"], "format": "date-time" },
          "lastSeenAt": { "type": "string", "format": "date-time" },
          "isFeatured": { "type": "boolean" },
          "freshness": { "$ref": "#/components/schemas/Freshness" },
          "industry": {
            "oneOf": [
              {
                "type": "object",
                "properties": {
                  "id": { "type": "string" },
                  "slug": { "type": "string" },
                  "name": { "type": "string" }
                },
                "required": ["id", "slug", "name"]
              },
              { "type": "null" }
            ]
          },
          "taxonomy": {
            "oneOf": [
              {
                "type": "object",
                "properties": {
                  "id": { "type": "string" },
                  "slug": { "type": "string" },
                  "canonicalTitle": { "type": "string" }
                },
                "required": ["id", "slug", "canonicalTitle"]
              },
              { "type": "null" }
            ]
          }
        },
        "required": ["id", "title", "company", "isRemote", "lastSeenAt", "isFeatured", "freshness"]
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "per_page": { "type": "integer" },
          "total": { "type": "integer" },
          "total_pages": { "type": "integer" },
          "has_next": { "type": "boolean" },
          "has_prev": { "type": "boolean" }
        },
        "required": ["page", "per_page", "total", "total_pages", "has_next", "has_prev"]
      },
      "JobSearchResponse": {
        "type": "object",
        "properties": {
          "jobs": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Job" }
          },
          "pagination": { "$ref": "#/components/schemas/Pagination" },
          "meta": {
            "type": "object",
            "properties": {
              "query_expanded_to": { "type": ["string", "null"] },
              "sort": { "type": "string" },
              "filters": {
                "type": "object",
                "properties": {
                  "role": { "type": ["string", "null"] },
                  "industry": { "type": ["string", "null"] },
                  "remote": { "type": ["boolean", "null"] },
                  "region": { "type": ["string", "null"] },
                  "posted_after": { "type": ["string", "null"] }
                }
              }
            }
          }
        },
        "required": ["jobs", "pagination", "meta"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "details": {}
        },
        "required": ["error"]
      }
    }
  },
  "security": [
    {},
    { "ApiKeyAuth": [] }
  ]
}